+++ capi/CHANGELOG.md
@@ -1,3 +1,8 @@
+Unreleased
+----------
+- Added `BLAZE_{NORMALIZE,SYMBOLIZE}_REASON_IGNORED_ERROR` constants
+
+
 0.1.5
 -----
 - Bumped `blazesym` dependency to `0.2.0`